pyproject.toml
[project]
name = "asgn2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.23.0",
    "pypdf2>=3.0.1",
]
